+++ pages/api/rooms/[id].ts
