+++ pages/api/rooms/index.ts
