+++ pages/api/rooms/join/[id].ts
