+++ pages/api/rooms/leave/[id].ts
